refactor: update package structure and rename category utility methods
